app/requirements.txt
asgiref==3.8.1
aspectlib==2.0.0
black==24.4.2
click==8.1.7
coverage==7.6.0
Django==5.0.7
django-extensions==3.2.3
django-timestampable==1.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==26.0.0
fields==5.0.0
flake8==7.1.0
freezegun==1.5.1
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-faker==2.0.0
python-dateutil==2.9.0.post0
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
wheel==0.43.0
